[#118] Use the <cstdint> macros for minimum-width integer constants (since e.g. uint64_t is unsigned long and requires UL suffix on some platforms and is unsigned long long and requires ULL on others (thanks to @ncpenke, Chandra Penke, Nextera Video, for the report)
(cherry picked from commit b87c1046b4a4a4e7c18862b95b28b103f3f6c12a)

+++ Development/rql/rql.cpp
@@ -652,7 +652,7 @@ namespace rql
             {
                 return value_indeterminate;
             }
-            return web::json::value::number(arg.size());
+            return web::json::value::number((uint64_t)arg.size());
         }
 
         // get(<call-operator>) - Value of the property specified by the result of the call-operator (rather than that value directly)
